# libs/api-types/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soroban Block Explorer API",
    "description": "REST API exposing ledger, transaction, contract, and NFT data for the Soroban block explorer. All list endpoints use cursor-based pagination (see ADR 0008).",
    "contact": {
      "name": "Rumble Fish",
      "url": "https://rumblefish.dev"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "tags": ["accounts"],
        "summary": "List accounts ordered by `last_seen_ledger` (the only indexed sort) —\nnewest-active first by default, oldest-first with `?order=asc`. The order\nis sticky across pages; cursor pagination walks within it.\n`filter[with_domain]` keeps only accounts that set a home_domain. No\naddress search — exact lookup is the global search's redirect path. Same\nshape as the other list endpoints.",
        "operationId": "list_accounts",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (1–100, default 20).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque pagination cursor from a previous response.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Base sort order on `last_seen_ledger`: `asc` | `desc` (default).\nSticky — re-send on every page alongside `cursor`.",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "name": "filter[with_domain]",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["boolean", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated account list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_AccountListItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}": {
      "get": {
        "tags": ["accounts"],
        "summary": "Account detail — header from `accounts` + balances from\n`account_balances_current` (canonical 06 statements A + B).",
        "operationId": "get_account",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Stellar account StrKey (G…, 56 chars)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account detail with current balances",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid account_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}/transactions": {
      "get": {
        "tags": ["accounts"],
        "summary": "Paginated transactions involving the account (source or participant).\n404 when the StrKey is unknown — distinct from \"indexed account, no\ntransactions yet\" (matches assets/contracts sub-resource pattern).",
        "operationId": "list_account_transactions",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Stellar account StrKey (G…, 56 chars)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (1–100, default 20).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque pagination cursor from a previous response.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order on transaction time: `asc` | `desc` (default).\nSticky — re-send on every page alongside `cursor`.",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated transactions involving the account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_AccountTransactionItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid account_id / pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountDetailResponse": {
        "type": "object",
        "required": [
          "account_id",
          "sequence_number",
          "balances",
          "first_seen_ledger",
          "last_seen_ledger",
          "deleted"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "balances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountBalance"
            }
          },
          "deleted": {
            "type": "boolean",
            "description": "`true` when the account was removed from the ledger via `account_merge`\nand never re-funded (its last lifecycle event is the merge). Derived,\nnot stored."
          },
          "first_seen_ledger": {
            "type": "integer",
            "format": "int64"
          },
          "home_domain": {
            "type": ["string", "null"]
          },
          "last_seen_ledger": {
            "type": "integer",
            "format": "int64"
          },
          "sequence_number": {
            "type": "integer",
            "format": "int64"
          },
          "signing": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AccountSigning",
                "description": "Signing configuration, or `null` when we have never observed this\naccount's entry state.\n\n**`null` is NOT \"no extra signers\".** 25% of indexed accounts carry no\nrow, so a client that renders `null` as \"single signature\" is making a\nsecurity claim the data does not support. Render the unknown as\nunknown."
              }
            ]
          }
        }
      },
...
      "AccountSigner": {
        "type": "object",
        "description": "One entry of an account's signer list, exactly as the ledger stores it.",
        "required": ["key", "weight", "type"],
        "properties": {
          "key": {
            "type": "string",
            "description": "The signer's StrKey. `ed25519` signers are `G…`; the hash types carry\ntheir own prefixes."
          },
          "type": {
            "type": "string",
            "description": "`ed25519` | `preauth_tx` | `hash_x` (the only three the network has)."
          },
          "weight": {
            "type": "integer",
            "format": "int32",
            "description": "1–255 for a real signer (`SetOptions` deletes at 0). Stored as the\nchain carried it — an out-of-range value is an anomaly, never clamped.",
            "minimum": 0
          }
        }
      },
      "AccountSigning": {
        "type": "object",
        "description": "An account's signing configuration: who may sign, and how much weight each\noperation class demands.\n\n**The master key is NOT in `signers`.** The ledger keeps the account's own\nkey out of that list and expresses its weight as `master_weight`; Horizon\nsynthesises an extra entry, we do not. A client showing only `signers`\ntherefore reads 3-of-4 where the chain says 3-of-5 — compose the master key\ninto the display, do not expect it here.\n\nA `master_weight` of 0 is a permanently DISABLED master key, not a signer\nwith a low weight. It is a common, deliberate configuration.",
        "required": [
          "signers",
          "master_weight",
          "threshold_low",
          "threshold_med",
          "threshold_high",
          "last_updated_ledger"
        ],
        "properties": {
          "last_updated_ledger": {
            "type": "integer",
            "format": "int64",
            "description": "Ledger at which this configuration was last observed."
          },
          "master_weight": {
            "type": "integer",
            "format": "int32",
            "description": "Weight of the account's own key — compare thresholds against\n`master_weight` PLUS the sum of `signers[].weight`.",
            "minimum": 0
          },
          "signers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountSigner"
            }
          },
          "threshold_high": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "threshold_low": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "threshold_med": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "description": "Canonical error response body.\n\nIntentionally simpler than RFC 7807 (`problem+json`) — see ADR 0008.\nMachine-readable `code` drives client behaviour; `message` is a\nhuman-friendly description; `details` carries optional structured\ncontext (field-level validation errors, etc.).",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code, e.g. `\"invalid_cursor\"`."
          },
          "details": {
            "description": "Optional structured details. Shape is error-specific."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error description."
          }
        }
      },
...
      "PageInfo": {
        "type": "object",
        "description": "Pagination metadata attached to list responses.\n\nCursor-based pagination (not offset) — see ADR 0008. Aligns with\nStellar Horizon conventions and produces stable listings even when\nthe underlying ledger stream advances between requests.\n\nBoth `next_cursor` (forward / next page) and `prev_cursor` (backward\n/ previous page) are **always emitted** as opaque base64-JSON\nstrings or explicit JSON `null`. Wire shape always carries both\nkeys — never omitted — so the OpenAPI `nullable: true` declaration\nmatches the on-wire bytes exactly. `Option<String>` carries the\nentire signal:\n\n- `next_cursor: null` ⇒ last page (no further forward continuation).\n- `next_cursor: \"...\"` ⇒ a further forward page exists.\n- `prev_cursor: null` ⇒ first page (no backward continuation).\n- `prev_cursor: \"...\"` ⇒ a backward page exists.\n\nBoth cursors carry an internal `dir` discriminant (see\n`crates/api/src/common/cursor.rs::Direction`) so the backend can\nbranch SQL between forward (DESC `<`) and backward (ASC `>`,\nreverse for presentation) walks. Clients treat the strings as\nopaque tokens — round-tripping them through `?cursor=` is the only\nexpected use.",
        "required": ["limit"],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32",
            "description": "Page size that produced `data`. Echoes the client's requested\nlimit (clamped server-side).",
            "maximum": 100,
            "minimum": 1
          },
          "next_cursor": {
            "type": ["string", "null"],
            "description": "Opaque cursor identifying the next page. `null` when the\nclient has reached the end of the stream."
          },
          "prev_cursor": {
            "type": ["string", "null"],
            "description": "Opaque cursor identifying the previous page. `null` when the\nresponse is the first page (the client has not paged forward\nyet)."
          }
        }
      },
      "Paginated_AccountListItem": {
        "type": "object",
        "description": "Canonical envelope for paginated list responses.\n\nGeneric over the item type `T` so every endpoint can reuse a single\nshape. Concrete instantiations (e.g. `Paginated<Transaction>`) are\npicked up automatically by utoipa-axum via the handler return type\nwhen M2 endpoint modules are wired in. Unused in M1 — kept as\ninfrastructure that M2 endpoints will consume.",
        "required": ["data", "page"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "One row of `GET /v1/accounts`. Identity + native (XLM) balance + the\nfirst/last-seen activity window + `home_domain`. Ordered by\n`last_seen_ledger` (the only indexed sort). `xlm_balance` is the native\nbalance from the unified `balances` table; `null` if no native row exists.",
              "required": [
                "account_id",
                "last_seen_ledger",
                "first_seen_ledger"
              ],
              "properties": {
                "account_id": {
                  "type": "string"
                },
                "first_seen_ledger": {
                  "type": "integer",
                  "format": "int64"
                },
                "home_domain": {
                  "type": ["string", "null"]
                },
                "last_seen_ledger": {
                  "type": "integer",
                  "format": "int64"
                },
                "xlm_balance": {
                  "type": ["string", "null"],
                  "description": "Native (XLM) balance as a RAW `Int128` stroop string (human value =\n÷10⁷); the frontend scales by 7 decimals — same raw-amount contract as\nthe account-detail balances."
                }
              }
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
      "Paginated_AccountTransactionItem": {
        "type": "object",
        "description": "Canonical envelope for paginated list responses.\n\nGeneric over the item type `T` so every endpoint can reuse a single\nshape. Concrete instantiations (e.g. `Paginated<Transaction>`) are\npicked up automatically by utoipa-axum via the handler return type\nwhen M2 endpoint modules are wired in. Unused in M1 — kept as\ninfrastructure that M2 endpoints will consume.",
        "required": ["data", "page"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Slim — `inner_tx_hash` lives on `/v1/transactions` only.",
              "required": [
                "hash",
                "ledger_sequence",
                "application_order",
                "source_account",
                "fee_charged",
                "successful",
                "operation_count",
                "has_soroban",
                "operation_types",
                "created_at"
              ],
              "properties": {
                "application_order": {
                  "type": "integer",
                  "format": "int32",
                  "description": "1-based position in ledger."
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "fee_charged": {
                  "type": "integer",
                  "format": "int64",
                  "description": "Fee charged, in raw stroops. Native (XLM) is always 7 decimals, so\nthere is no `decimals` field — the frontend scales by 1e7."
                },
                "has_soroban": {
                  "type": "boolean"
                },
                "hash": {
                  "type": "string",
                  "description": "64-char lowercase hex."
                },
                "ledger_sequence": {
                  "type": "integer",
                  "format": "int64"
                },
                "operation_count": {
                  "type": "integer",
                  "format": "int32"
                },
                "operation_types": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "source_account": {
                  "type": "string"
                },
                "successful": {
                  "type": "boolean"
                }
              }
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "bearer_jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "api_key": []
    },
    {
      "bearer_jwt": []
    }
  ]
}